JAMES-3400 Move HTTP status codes to a constant
